Update README.md
In the 2nd project at Istanbul Data Science Bootcamp, I applied Linear Regression, Lasso and Ridge Regression Techniques to the data was scraped from IMDB.com and determined the important features that affect a movie. When a new movie is released, I determined the most outperformed model to predict the international revenue of this movie.